Update some files.

+++ Symfony iOS Oauth Demo/OAuth2Client/Sources/OAuth2Client/NXOAuth2PostBodyStream.m
@@ -88,7 +88,7 @@
             NSData *delimiterData = [delimiter dataUsingEncoding:NSUTF8StringEncoding];
             NSData *contentHeaderData = [[part contentHeaders] dataUsingEncoding:NSUTF8StringEncoding];
             
-            int dataLength = delimiterData.length + contentHeaderData.length;
+            long dataLength = delimiterData.length + contentHeaderData.length;
             NSMutableData *headerData = [NSMutableData dataWithCapacity: dataLength];
             [headerData appendData:delimiterData];
             [headerData appendData:contentHeaderData];
@@ -146,7 +146,7 @@
     if (currentStream == nil)
         return 0;
     
-    int result = [currentStream read:buffer maxLength:len];
+    long result = [currentStream read:buffer maxLength:len];
     
     if (result == 0) {
         if (streamIndex < contentStreams.count - 1) {
